move implementation of BasicQueryableMemoryDB into separate file
Make it easier to find the implementation of the DB used for tests
